Start reworking interaction between chips
This commit lays the groundwork for cycle accurate emulation. The PPU is
now no longer being stepped in together with the CPU but instead it is
not stepped at all unless the CPU tries to interact with it, in which
case it catches up with the CPU before handling the interaction.

The crucial difference here is that we can make the PPU catch up with
the CPU in the middle of executing an instruction whereas before the CPU
interpreters step function had to return before the other chips could be
stepped.